Split embedding ANN index module
